pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poe.tasks]
dev = "streamlit run --server.runOnSave=true --server.headless=true streamlit_app.py"
run = "streamlit run streamlit_app.py"

[tool.poetry]
authors = ["Matt Sárdi <sardi.matt@gmail.com>"]
description = ""
name = "simplex-2021-demo"
version = "0.1.0"

[tool.poetry.dependencies]
Simplex = {git = "https://github.com/gmatt/Simplex.git", rev="eb7436d6d90b02eb875d2bb6952ccc754b221e34"}
matplotlib = "^3.5.1"
openTSNE = "^0.6.1"
pandas = "^1.3.3"
pyprojroot = "^0.2.0"
python = ">=3.8,<3.10"
scikit-learn = "^0.24.2"
seaborn = "0.11.2"
streamlit = "^1.3.1"
streamlit-ace = "^0.1.1"
streamlit-plotly-events = "^0.0.6"
torch = "^1.9.1"
umap-learn = "^0.5.2"

[tool.poetry.dev-dependencies]
poethepoet = "^0.11.0"
pytest = "^5.2"
watchdog = "^2.1.6"
